+++ libraries/provider_chef_dk_mac_os_x.rb
@@ -70,6 +70,15 @@ class Chef
         def package_file_extension
           '.dmg'
         end
+
+        #
+        # Return the global bashrc file path for this system
+        #
+        # @return [String]
+        #
+        def bashrc_file
+          '/etc/bashrc'
+        end
       end
     end
   end